Fix field filtering and default values
Restores capability removed in https://github.com/UKHomeOfficeForms/hof-behaviour-summary-page/pull/4 for a default value to be used in place of undefined or empty values.

Also fixes functionality claimed to be added by that PR of not filtering zero-valued fields, which was broken due to the secondary pure-Boolean filter that was applied after the initial filter.

Now fields are only removed if they have a falsy, non-zero value *and* no default value is supplied.